+++ src/components/article/Hero.tsx
@@ -3,7 +3,7 @@ import styled from "styled-components";
 import { LeadText } from "../basic/Typography";
 import { HeaderWithBackground } from "../layout/Header";
 import React from "react";
-import PostMeta from "../blog/meta/PostMeta";
+import PostMeta from "./meta/PostMeta";
 import { HeroTitle } from "../layout/Hero";
 
 const Lead = styled(LeadText)``;
